+++ lib/floe-core/default.nix
@@ -6,17 +6,17 @@
 
 let
   types = import ./types.nix { inherit lib; };
-  interfaces = import ./interfaces.nix { inherit lib; };
+  interfaces = import ./interfaces.nix { inherit lib types; };
   floe = import ./floe.nix { inherit lib types; };
   link = import ./link.nix {
-    inherit lib types;
+    inherit lib types interfaces;
     floeLib = floe;
   };
 in
 {
   T = types;
   inherit (types) checkValue isDeferredToken;
-  inherit (interfaces) mkSig mkOutputKind;
+  inherit (interfaces) mkSig mkOutputKind isUncrossable;
   inherit (floe) mkFloe;
   inherit (link) link;
 }